rename HIRT_Multichannel_Convolution/{ConvolvePartitioned => partitioned}.hpp
